# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 'Test title for docs'
  description: ''
  version: 1.0.0
servers:
  -
    url: 'http://localhost'
paths:
...
  '/api/v1/tickets/{id}':
    get:
      summary: show
      operationId: show
      description: '<example description>'
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  data: 'Example success response'
                properties:
                  data:
                    type: string
                    example: 'Example success response'
        422:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  errors: 'Example error response'
                properties:
                  errors:
                    type: string
                    example: 'Example error response'
      tags:
        - Tickets
    put:
      summary: update
      operationId: update
      description: '<example description>'
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  data: 'Example success response'
                properties:
                  data:
                    type: string
                    example: 'Example success response'
        422:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  errors: 'Example error response'
                properties:
                  errors:
                    type: string
                    example: 'Example error response'
      tags:
        - Tickets
    delete:
      summary: destroy
      operationId: destroy
      description: '<example description>'
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  data: 'Example success response'
                properties:
                  data:
                    type: string
                    example: 'Example success response'
        422:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  errors: 'Example error response'
                properties:
                  errors:
                    type: string
                    example: 'Example error response'
      tags:
        - Tickets
    parameters:
      -
        in: path
        name: id
        description: 'The ID of the ticket.'
        example: 1
        required: true
        schema:
          type: integer
  '/api/v1/tickets/{ticket':
    put:
      summary: replace
      operationId: replace
      description: '<example description>'
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  data: 'Example success response'
                properties:
                  data:
                    type: string
                    example: 'Example success response'
        422:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  errors: 'Example error response'
                properties:
                  errors:
                    type: string
                    example: 'Example error response'
      tags:
        - Tickets
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                data:
                  type: object
                  description: ''
                  example: []
                  properties:
                    attributes:
                      type: object
                      description: ''
                      example: []
                      properties:
                        title:
                          type: string
                          description: ''
                          example: harum
                        description:
                          type: string
                          description: ''
                          example: quibusdam
                        status:
                          type: string
                          description: ''
                          example: X
                          enum:
                            - A
                            - C
                            - H
                            - X
                    relationships:
                      type: object
                      description: ''
                      example: []
                      properties:
                        author:
                          type: object
                          description: ''
                          example: []
                          properties:
                            data:
                              type: object
                              description: ''
                              example: []
                              properties:
                                id:
                                  type: integer
                                  description: ''
                                  example: 4
...
  '/api/v1/users/{id}':
    get:
      summary: show
      operationId: show
      description: '<example description>'
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  data: 'Example success response'
                properties:
                  data:
                    type: string
                    example: 'Example success response'
        422:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  errors: 'Example error response'
                properties:
                  errors:
                    type: string
                    example: 'Example error response'
      tags:
        - Users
    put:
      summary: update
      operationId: update
      description: '<example description>'
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  data: 'Example success response'
                properties:
                  data:
                    type: string
                    example: 'Example success response'
        422:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  errors: 'Example error response'
                properties:
                  errors:
                    type: string
                    example: 'Example error response'
      tags:
        - Users
    delete:
      summary: destroy
      operationId: destroy
      description: '<example description>'
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  data: 'Example success response'
                properties:
                  data:
                    type: string
                    example: 'Example success response'
        422:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  errors: 'Example error response'
                properties:
                  errors:
                    type: string
                    example: 'Example error response'
      tags:
        - Users
    parameters:
      -
        in: path
        name: id
        description: 'The ID of the user.'
        example: 1
        required: true
        schema:
          type: integer
...
components:
  securitySchemes:
    default:
      type: http
      scheme: bearer
      description: 'You can retrieve your token by visiting your dashboard and clicking <b>Generate API token</b>.'
security:
  -
    default: []